+++ src/app/ApplicationTemplate.Access/Framework/Configuration/IConfiguration.Extensions.cs
@@ -14,7 +14,7 @@ public static class ApplicationTemplateConfigurationExtensions
 {
 	/// <summary>
 	/// Returns the <see cref="IConfigurationSection"/> for the options represented by <typeparamref name="T"/>.
-	/// The section name is either the options type name (minus the -Options prefix) or <paramref name="key"/>, if provided.
+	/// The section name is either the options type name (minus the -Options suffix) or <paramref name="key"/>, if provided.
 	/// <see cref="DefaultOptionsName{T}"/> as well.
 	/// </summary>
 	/// <typeparam name="T">The options type.</typeparam>
@@ -28,7 +28,7 @@ public static class ApplicationTemplateConfigurationExtensions
 
 	/// <summary>
 	/// Reads the current value for options.
-	/// The section name is either the options type name (minus the -Options prefix) or <paramref name="key"/>, if provided.
+	/// The section name is either the options type name (minus the -Options suffix) or <paramref name="key"/>, if provided.
 	/// <see cref="DefaultOptionsName{T}"/> as well.
 	/// </summary>
 	/// <typeparam name="T">The options type.</typeparam>
@@ -44,19 +44,21 @@ public static class ApplicationTemplateConfigurationExtensions
 
 	/// <summary>
 	/// Gets the default name for options of type <paramref name="optionsType"/>.
-	/// Removes the -Options prefix if any.
+	/// Removes the -Options suffix if any.
 	/// </summary>
+	/// <param name="optionsType">The type from which to extract a default options name.</param>
 	public static string DefaultOptionsName(Type optionsType) => DefaultOptionsName(optionsType.Name);
 
 	/// <summary>
 	/// Gets the default name for options of the specified <paramref name="optionsTypeName"/>.
-	/// Removes the -Options prefix if any.
+	/// Removes the -Options suffix if any.
 	/// </summary>
+	/// <param name="optionsTypeName">The name of the type from which to extract a default options name.</param>
 	public static string DefaultOptionsName(string optionsTypeName) => Regex.Replace(optionsTypeName, @"Options$", string.Empty);
 
 	/// <summary>
 	/// Gets the default name for options of type <typeparamref name="T"/>.
-	/// Removes the -Options prefix if any.
+	/// Removes the -Options suffix if any.
 	/// </summary>
 	/// <typeparam name="T">The Options type.</typeparam>
 	public static string DefaultOptionsName<T>() => DefaultOptionsName(typeof(T));