+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 Prefix your items with `(Template)` if the change is about the template and not the resulting application.
 
+## 2.1.X
+- Enable `TreatWarningsAsErrors` for the Access, Business, and Presentation projects.
+- Update analyzers packages and severity of rules.
+
 ## 2.0.X
 - Renamed the classes providing data to use the `Repository` suffix instead of `Endpoint` or `Service`.
 - Renamed the Client library to Access.